% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsense.R
\name{getD}
\alias{getD}
\title{Auxiliary function for getFC()}
\usage{
getD(value, FC, thFoldChange = NA, pValueFC = 0.05)
}
\arguments{
\item{dataset}{data.frame, rows correspond to different genes, first column contains gene identifiers, second column contains the gene name, columns 3 to n contain the RNAseq count data, column names should start with the condition identifier (e.g. "WT") followed by the time separated by "_"}

\item{myanalyzeConditions}{Character vector, Name of experimental conditions}

\item{cores}{Numeric, Number of cores for parallelization, default 1 for no parallelization}

\item{mytimes}{Numeric vector, Time points of the time-resolved RNA-seq data}
}
\value{
Data.frame containing gene names, log fold change and p-values calculated from NBPSeq, each gene appears as often as available time points
}
\description{
Auxiliary function for getFC()
}
\author{
Marcus Rosenblatt, \email{marcus.rosenblatt@fdm.uni-freiburg.de}
}
