% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MZsox}
\alias{MZsox}
\title{Time resolved RNA seq data for zebra fish MZsox and wild-type.}
\format{A SummarizedExperiment object with 15775 rows and 40 columns:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
MZsox
}
\description{
A dataset containing RNAseq measurements for two experimental conditions at 7 different time points including replicates.
}
\keyword{datasets}
