% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsense.R
\name{plotSSGS}
\alias{plotSSGS}
\title{plot SSGS gene classes}
\usage{
plotSSGS(myresultCombined = resultCombined, mytimes = times,
  myanalyzeConditions = analyzeConditions)
}
\arguments{
\item{myresultCombined}{data.frame, output of \link{combineResults}}

\item{mytimes}{Numeric vector, Time points of the time-resolved RNA-seq data}

\item{myanalyzeConditions}{character vector, the conditions that were analyzed}
}
\value{
SSGS color plot in ggplot format
}
\description{
Genes are sorted into groups with respect to switch time and time point of fold change detection. For each group, results of wild type and knockdown-condition are compared by means of fisher's exact test to show whether the knocked down gene enhances or suppresses the respective gene group.
}
\author{
Marcus Rosenblatt, \email{marcus.rosenblatt@fdm.uni-freiburg.de}
}
