% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsense.R
\name{combineResults}
\alias{combineResults}
\title{Combine results}
\usage{
combineResults(myresultSwitch = resultSwitch, myresultFC = resultFC)
}
\arguments{
\item{myresultSwitch}{data.frame, output of getSwitch}

\item{myresultFC}{data.frame, output of getFC}
}
\value{
Data.frame containing information on switch and fold change detection for each gene
}
\description{
Results of switch and fold change analysis are collected in one data.frame
}
\author{
Marcus Rosenblatt, \email{marcus.rosenblatt@fdm.uni-freiburg.de}
}
