% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsense.R
\name{getD}
\alias{getD}
\title{Auxiliary function for getFC()}
\usage{
getD(value, FC, thFoldChange = NA, pValueFC = 0.05)
}
\arguments{
\item{value}{output pValue from exact.nb.test}

\item{FC}{output logFoldChange from exact.nb.test}

\item{thFoldChange}{Numeric, threshold at which fold change is counted to be detected}

\item{pValueFC}{Numeric, p-value at which fold change is counted to be detected}
}
\value{
input plus decision whether fold change has been detected or not
}
\description{
Auxiliary function for getFC()
}
\author{
Marcus Rosenblatt, \email{marcus.rosenblatt@fdm.uni-freiburg.de}
}
