% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsense.R
\name{getSwitch}
\alias{getSwitch}
\title{Detect switching genes}
\usage{
getSwitch(dataset = mydata, experimentStepDetection = "WT",
  pValueSwitch = 0.05, cores = 1)
}
\arguments{
\item{dataset}{data.frame, rows correspond to different genes, first column contains gene identifiers, second column contains the gene name, columns 3 to n contain the RNAseq count data, column names should start with the condition identifier (e.g. "WT") followed by the time separated by "_"}

\item{experimentStepDetection}{Character, Name of condition for which switch detection is performed}

\item{pValueSwitch}{Numeric, A threshold for counting cells as being invaded or not. When cells move towards negative z-direction, threshold should be negative.}

\item{cores}{Numeric, Number of cores for parallelization, default 1 for no parallelization}
}
\value{
Data.frame containing gene names and results of switch detection, information about switch time point and direction
}
\description{
For each gene, time-resolved RNA-seq measurements are analyzed for occurence of switches (up or down)
}
\author{
Marcus Rosenblatt, \email{marcus.rosenblatt@fdm.uni-freiburg.de}
}
