% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAsense.R
\name{outputGeneTables}
\alias{outputGeneTables}
\title{Output gene tables}
\usage{
outputGeneTables(myresultCombined = resultCombined, mytimes = times)
}
\arguments{
\item{myresultCombined}{data.frame, output of combineResults}

\item{mytimes}{numeric vector, the measurement times}
}
\value{
current working directory
}
\description{
Output information on switching genes (up/down) in tabular format (gene identifier/gene name) are created as csv file and written to the current working directory
}
\author{
Marcus Rosenblatt, \email{marcus.rosenblatt@fdm.uni-freiburg.de}
}
